fix: quickFix's parameter type & command name handling (#200)

+++ jdtls.ext/com.shengchen.checkstyle.runner/src/main/java/com/shengchen/checkstyle/runner/CheckstyleRunner.java
@@ -97,7 +97,7 @@ public class CheckstyleRunner {
 
     public static WorkspaceEdit quickFix(
         String fileToCheckUri,
-        int offset,
+        Double offset,
         String sourceName
     ) throws JavaModelException, IllegalArgumentException, BadLocationException {
         final BaseQuickFix quickFix = QuickFixProvider.getQuickFix(sourceName);
@@ -107,13 +107,13 @@ public class CheckstyleRunner {
 
         final ICompilationUnit unit = JDTUtils.resolveCompilationUnit(fileToCheckUri);
         final Document document = new Document(unit.getSource());
-        final IRegion lineInfo = document.getLineInformationOfOffset(offset);
+        final IRegion lineInfo = document.getLineInformationOfOffset(offset.intValue());
         final ASTParser astParser = ASTParser.newParser(IASTSharedValues.SHARED_AST_LEVEL);
         astParser.setKind(ASTParser.K_COMPILATION_UNIT);
         astParser.setSource(unit);
         final CompilationUnit astRoot = (CompilationUnit) astParser.createAST(null);
         astRoot.recordModifications();
-        astRoot.accept(quickFix.getCorrectingASTVisitor(lineInfo, offset));
+        astRoot.accept(quickFix.getCorrectingASTVisitor(lineInfo, offset.intValue()));
         final TextEdit edit = astRoot.rewrite(document, null);
         return EditUtils.convertToWorkspaceEdit(unit, edit);
     }
